fixing ide controller
