+++ elenasrc3/ide/ideversion.h
@@ -1,2 +1,2 @@
 
-#define IDE_REVISION_NUMBER                           1
+#define IDE_REVISION_NUMBER                           2
